requirements.txt
opencv-python-headless==4.6.0.66
pytesseract==0.3.7
numpy==1.23.2
Flask==2.1.0
Pillow==9.3.0
